Created separate models for system bodies and stations
