+++ taxee_server/taxee_server/Models/Body.cs
@@ -1,0 +1,9 @@
+﻿namespace taxee_server.Models
+{
+    public class Body
+    {
+        public string BodyID { get; set; }
+        public string BodyName { get; set; }
+        public int SystemID { get; set; }
+    }
+}
