+++ taxee_server/taxee_server/Models/Station.cs
@@ -1,0 +1,9 @@
+﻿namespace taxee_server.Models
+{
+    public class Station
+    {
+        public string StationID { get; set; }
+        public string StationName { get; set; }
+        public int SystemID { get; set; }
+    }
+}
